Windows fixes
Maintaing a unix like software for windows is a curse ...

+++ pwman/tests/test_pwman.py
@@ -25,10 +25,12 @@ from .db_tests import (DBTests, SetupTester, CLITests, ConfigTest,
                        TestDBFalseConfig, FactoryTest)
 
 from .crypto_tests import CryptoTest
-from .test_complete_ui import Ferrum, NEW_DB_PATH
 
-if os.path.exists(NEW_DB_PATH):
-    os.remove(NEW_DB_PATH)
+if 'win' not in sys.platform:
+    from .test_complete_ui import Ferrum, NEW_DB_PATH
+
+    if os.path.exists(NEW_DB_PATH):
+        os.remove(NEW_DB_PATH)
 
 # make sure we use local pwman
 sys.path.insert(0, os.getcwd())
@@ -43,11 +45,12 @@ def suite():
     suite = unittest.TestSuite()
     suite.addTest(loader.loadTestsFromTestCase(DBTests))
     suite.addTest(loader.loadTestsFromTestCase(CryptoTest))
-    #suite.addTest(loader.loadTestsFromTestCase(CLITests))
-    #suite.addTest(loader.loadTestsFromTestCase(ConfigTest))
-    #suite.addTest(loader.loadTestsFromTestCase(FactoryTest))
-    #suite.addTest(loader.loadTestsFromTestCase(TestDBFalseConfig))
-    #suite.addTest(loader.loadTestsFromTestCase(Ferrum))
+    suite.addTest(loader.loadTestsFromTestCase(CLITests))
+    suite.addTest(loader.loadTestsFromTestCase(ConfigTest))
+    suite.addTest(loader.loadTestsFromTestCase(FactoryTest))
+    suite.addTest(loader.loadTestsFromTestCase(TestDBFalseConfig))
+    if 'win' not in sys.platform:
+        suite.addTest(loader.loadTestsFromTestCase(Ferrum))
     return suite
 
 if __name__ == '__main__':
